+++ SaluteStocksAPI/Models/FundamentalData/Earnings.cs
@@ -30,7 +30,7 @@ public class AnnualEarning
     /// привлекательности и эффективности компаний,
     /// действующих на фондовом рынке.
     /// </summary>
-    [JsonProperty("reportedEPS")] public decimal? ReportedEPS { get; set; }
+    [JsonProperty("reportedEPS")] public double? ReportedEPS { get; set; }
 }
 
 public class QuarterlyEarning
@@ -46,13 +46,13 @@ public class QuarterlyEarning
 
     [JsonProperty("reportedDate")] public DateTime? ReportedDate { get; set; }
 
-    [JsonProperty("reportedEPS")] public decimal? ReportedEPS { get; set; }
+    [JsonProperty("reportedEPS")] public double? ReportedEPS { get; set; }
 
-    [JsonProperty("estimatedEPS")] public decimal? EstimatedEPS { get; set; }
+    [JsonProperty("estimatedEPS")] public double? EstimatedEPS { get; set; }
 
-    [JsonProperty("surprise")] public decimal? Surprise { get; set; }
+    [JsonProperty("surprise")] public double? Surprise { get; set; }
 
-    [JsonProperty("surprisePercentage")] public decimal? SurprisePercentage { get; set; }
+    [JsonProperty("surprisePercentage")] public double? SurprisePercentage { get; set; }
 }
 
 public class Earnings : CompanyEntityInfo
